fixing bugs for multifile torrents and torrents > 2GB
git-svn-id: http://svn.vuze.com/public/client@11581 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/ui/swt/views/tableitems/pieces/BlocksItem.java
@@ -160,9 +160,9 @@ public class BlocksItem
       int pieceNumber = pePiece.getPieceNumber();
       long[] offsets = new long[(int)lNumBlocks];
       long[] lengths = (long[])offsets.clone();
-      Arrays.fill(offsets, pePiece.getManager().getDiskManager().getPieceLength()*pieceNumber);
+      Arrays.fill(offsets, (long)pePiece.getManager().getDiskManager().getPieceLength()*(long)pieceNumber);
       for (int i = 0; i<lNumBlocks;lengths[i]=pePiece.getBlockSize(i),offsets[i]+=DiskManager.BLOCK_SIZE * i,i++ );
-      
+
       boolean[] isCached = cacheStats == null ? null : cacheStats.getBytesInCache(torrent,offsets,lengths);
